requirements.txt
Flask==1.1.2
twilio==6.56.0
pandas==1.2.4
spacy==3.0.5
oauth2client==4.1.3
gspread==3.7.0
pycryptodome==3.10.1
pylint
autopep8